+++ Library/Aphid/Support.js
@@ -7,6 +7,7 @@
 
 Aphid.Support = {};
 
+//= require <Aphid/Support/Properties>
 //= require <Aphid/Support/Compatibility>
 //= require <Aphid/Support/Extensions>
 //= require <Aphid/Support/Cookie>