+++ CHANGELOG.markdown
@@ -3,6 +3,9 @@
 
 ## Version 0.9.2 - *Not Yet Released*
 
+ * [Aphid.Support] Added Aphid.Support.Properties, a mixin that defines a
+   standard accessor and setter pattern for Objects in Aphid.
+
  * [Tools] Added initial support for publishing the built project output of
    Aphid-based projects by using "rake publish" (publishing is configured in
    the Publish.rb file in the Aphid project root).